requirements/development.txt
click==6.6
coreapi==2.0.8
coreapi-cli==1.0.6
Django==1.10.2
django-autofixture==0.12.1
django-bootstrap-breadcrumbs==0.8
django-braces==1.9.0
django-ckeditor==5.1.1
django-filter==1.0.1
django-fontawesome==0.2.6
django-mptt==0.8.6
django-oauth-toolkit==0.10.0
djangorestframework==3.4.7
httpie==0.9.6
itypes==1.1.0
Markdown==2.6.7
MySQL-python==1.2.5
oauthlib==1.0.3
optional-django==0.3.0
Pygments==2.1.3
PyYAML==3.12
requests==2.11.1
six==1.10.0
uritemplate==3.0.0
